refactor: move and rename components to create class' content
