+++ src/app/(private)/courses/[id]/classes/[classId]/manage/_components/drag-drop-area.tsx
